@@ -1,53 +1,42 @@
 {
-  "name": "frontscope",
-  "version": "0.2.0",
-  "private": true,
-  "scripts": {
-    "serve": "vue-cli-service serve",
-    "build": "vue-cli-service build",
-    "test:vue": "vue-cli-service test:unit",
-    "lint": "vue-cli-service lint",
-    "format-check-ts": "prettier-eslint \"src/**/*.ts\"",
-    "format-write-ts": "prettier-eslint \"src/**/*.ts\" --write",
-    "format-check-vue": "prettier-eslint \"src/**/*.vue\"",
-    "format-write-vue": "prettier-eslint \"src/**/*.vue\" --write",
-    "format-check": "npm run format-check-ts && npm run format-check-vue",
-    "format-write": "npm run format-write-ts && npm run format-write-vue"
-  },
-  "dependencies": {
-    "axios": "^0.21.2",
-    "core-js": "^3.6.4",
-    "jquery": "^3.4.1",
-    "mathjs": "^9.5.1",
-    "p5": "^1.0.0",
-    "router": "^1.3.5",
-    "vue": "^2.6.11",
-    "vue-js-modal": "^2.0.0-rc.3",
-    "vue-router": "^3.1.6",
-    "vue-unique-id": "^3.1.0"
-  },
-  "devDependencies": {
-    "@types/jest": "^24.0.19",
-    "@types/p5": "^0.9.0",
-    "@typescript-eslint/eslint-plugin": "^4.32.0",
-    "@typescript-eslint/parser": "^4.32.0",
-    "@vue/cli-plugin-babel": "~4.3.0",
-    "@vue/cli-plugin-eslint": "~4.3.0",
-    "@vue/cli-plugin-router": "^4.2.3",
-    "@vue/cli-plugin-typescript": "~4.3.0",
-    "@vue/cli-plugin-unit-jest": "~4.5.0",
-    "@vue/cli-service": "~4.3.0",
-    "@vue/eslint-config-typescript": "^7.0.0",
-    "@vue/test-utils": "^1.3.0",
-    "css-loader": "^3.4.2",
-    "eslint": "^7.32.0",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-vue": "^7.18.0",
-    "prettier": "^2.5.1",
-    "prettier-eslint": "^13.0.0",
-    "prettier-eslint-cli": "^5.0.1",
-    "typescript": "~3.8.3",
-    "vue-loader": "^15.9.1",
-    "vue-template-compiler": "^2.6.11"
-  }
+    "name": "frontscope",
+    "version": "0.3.0",
+    "scripts": {
+        "dev": "vite",
+        "build": "vue-tsc --noEmit && vite build",
+        "preview": "vite preview --port 5050",
+        "test:unit": "vitest run",
+        "typecheck": "vue-tsc --noEmit -p tsconfig.vitest.json --composite false",
+        "lint": "prettier . --write && eslint . --ext .vue,.js,.jsx,.cjs,.mjs,.ts,.tsx,.cts,.mts --fix --ignore-path .gitignore",
+        "prepare": "husky install"
+    },
+    "dependencies": {
+        "p5": "^1.4.1",
+        "vue": "^3.2.29",
+        "vue-router": "^4.0.12"
+    },
+    "devDependencies": {
+        "@rushstack/eslint-patch": "^1.1.0",
+        "@types/jsdom": "^16.2.14",
+        "@types/node": "^16.11.22",
+        "@types/p5": "^1.4.2",
+        "@vitejs/plugin-vue": "^2.1.0",
+        "@vue/eslint-config-prettier": "^7.0.0",
+        "@vue/eslint-config-typescript": "^10.0.0",
+        "@vue/test-utils": "^2.0.0-rc.18",
+        "@vue/tsconfig": "^0.1.3",
+        "axios": "^0.27.2",
+        "eslint": "^8.5.0",
+        "eslint-plugin-vue": "^8.2.0",
+        "happy-dom": "^2.39.1",
+        "husky": "^7.0.4",
+        "jsdom": "^19.0.0",
+        "mathjs": "^10.5.3",
+        "prettier": "^2.5.1",
+        "typescript": "~4.5.5",
+        "vite": "^2.7.13",
+        "vitest": "^0.2.5",
+        "vue-tsc": "^0.31.1",
+        "vue-unique-id": "^3.2.1"
+    }
 }
