@@ -1,12 +1,9 @@
-import Vue from 'vue'
+import {createApp} from 'vue'
 import App from './App.vue'
 import router from './router'
-import UniqueId from 'vue-unique-id'
 
-Vue.config.productionTip = false
-Vue.use(UniqueId)
+const app = createApp(App)
 
-new Vue({
-    router,
-    render: h => h(App),
-}).$mount('#app')
+app.use(router)
+
+app.mount('#app')
