@@ -57,10 +57,8 @@
 </template>
 
 <script lang="ts">
-    import Vue from 'vue'
-    import BundleCard from '@/components/bundleCard.vue'
-
-    export default Vue.extend({
+    import BundleCard from './BundleCard.vue'
+    export default {
         components: {
             BundleCard,
         },
@@ -72,6 +70,7 @@
         computed: {
             readyToBundle: function (): boolean {
                 return (
+                    // prettier-ignore
                     this.activeSeq !== null
                     && this.activeSeq.isValid !== undefined
                     && this.activeViz !== null
@@ -79,5 +78,5 @@
                 )
             },
         },
-    })
+    }
 </script>
