@@ -11,7 +11,7 @@
     </div>
 </template>
 
-<script>
+<script lang="ts">
     export default {
         name: 'SeqSelector',
         props: {
@@ -25,12 +25,10 @@
     a {
         cursor: pointer;
     }
-
     li:hover {
         background-color: #809fff;
         color: #ffffff;
     }
-
     li {
         text-align: left;
     }
