@@ -10,7 +10,7 @@
     </div>
 </template>
 
-<script>
+<script lang="ts">
     export default {
         props: {
             buttonText: String,
