@@ -4,10 +4,9 @@
     </a>
 </template>
 
-<script>
+<script lang="ts">
     export default {
         name: 'VizSelector',
-
         props: {
             title: String,
         },
@@ -18,12 +17,10 @@
     a {
         cursor: pointer;
     }
-
     li:hover {
         background-color: #809fff;
         color: #ffffff;
     }
-
     li {
         text-align: left;
     }
