@@ -9,8 +9,9 @@ import {SequenceCached} from './SequenceCached'
  */
 class SequenceRandom extends SequenceCached {
     name = 'Random Integers in Range'
-    description =
-        'A sequence of integers chosen indepenently uniformly '
+    // prettier-ignore
+    description
+        = 'A sequence of integers chosen independently uniformly '
         + 'from n to m inclusive.'
     min = 0
     max = 9
