@@ -1,16 +1,20 @@
-import {VisualizerExportModule} from './VisualizerInterface'
-const vizFiles = require.context('.', false, /\.ts$/)
+import type {VisualizerExportModule} from './VisualizerInterface'
 
+// import each module in the current directory with a .ts file extension
+const vizFiles = import.meta.globEager('./*.ts')
+
+// object we will export
 const vizMODULES: {[key: string]: VisualizerExportModule} = {}
 
-const vizKeyArray: string[] = []
-vizFiles.keys().forEach(key => {
-    if (key === './index.js' || key === './modules.js') return
-    const exprt = vizFiles(key)
-    if (exprt.exportModule) {
-        vizKeyArray.push(key)
-        vizMODULES[key.replace(/(\.\/|\.ts)/g, '')] = exprt.exportModule
+// for each file in the list of visualizer files...
+for (const file in vizFiles) {
+    // if it has an export module...
+    if (vizFiles[file].exportModule) {
+        // add the export module to our object
+        // prettier-ignore
+        vizMODULES[file.replace(/(\.\/|\.ts)/g, '')]
+        = vizFiles[file].exportModule
     }
-})
+}
 
 export default vizMODULES
